requirements.txt
# Mutual Funds Agent Dependencies
# Core dependencies
aiohttp==3.9.5
asyncio-throttle==1.0.2
python-dotenv==1.0.0

# Data processing
pandas==2.2.2
numpy==1.26.4

# Web scraping
beautifulsoup4==4.12.3
lxml==5.2.2
requests==2.31.0

# LLM and AI
langchain==0.2.11
langchain-community==0.2.10
openai>=1.0.0

# Evaluation & Metrics
deepeval>=0.21.0
psycopg2-binary>=2.9.9
tabulate>=0.9.0

# Development and testing
pytest==8.2.2
pytest-asyncio==0.23.7
black==24.4.2
flake8==7.1.0

# Optional: FastAPI for API mode (future enhancement)
fastapi==0.111.0
uvicorn==0.30.1
websockets>=11.0.0
python-multipart>=0.0.6
pydantic>=2.0.0
